+++ mod/src/main/java/basemod/interfaces/ImGuiSubscriber.java
@@ -1,0 +1,5 @@
+package basemod.interfaces;
+
+public interface ImGuiSubscriber extends ISubscriber {
+	void receiveImGui();
+}